adjusted model reference to match file rename
