+++ server/server.js
@@ -8,7 +8,7 @@ const { typeDefs, resolvers } = require('./schemas');
 // //import Authentication middleware
 // const { authMiddleware } = require("./utils/auth");
 
-const db = require('./config/connections');
+const db = require('./config/connection');
 
 const PORT = process.env.PORT || 3001;
 const app = express();
